ObjectDetector/requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
filelock==3.8.2
h11==0.14.0
huggingface-hub==0.11.1
idna==3.4
numpy==1.24.0
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
pydantic==1.10.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
sniffio==1.3.0
starlette==0.22.0
timm==0.6.12
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
